+++ pysd/builders/python/python_functions.py
@@ -33,6 +33,10 @@ functionspace = {
     "sum": ("sum(%(0)s, dim=%(axis)s)", ("functions", "sum")),
     "vmax": ("vmax(%(0)s, dim=%(axis)s)", ("functions", "vmax")),
     "vmin": ("vmin(%(0)s, dim=%(axis)s)", ("functions", "vmin")),
+    "vector_select": (
+        "vector_select(%(0)s, %(1)s, %(axis)s, %(2)s, %(3)s, %(4)s)",
+        ("functions", "vector_select")
+    ),
 
     # functions defined in pysd.py_bakcend.functions
     "active_initial": (
@@ -63,6 +67,9 @@ functionspace = {
     "Xpulse_train": (
         "pulse(__data['time'], %(0)s, repeat_time=%(1)s, magnitude=%(2)s)",
         ("functions", "pulse")),
+    "get_time_value": (
+        "get_time_value(__data['time'], %(0)s, %(1)s, %(2)s)",
+        ("functions", "get_time_value")),
     "quantum": (
         "quantum(%(0)s, %(1)s)",
         ("functions", "quantum")),